UnitAura Now Works
UnitAura works for debuffs on non-self targets, it does not appear to work with buffs applied by other players to other players, but that is a very edge case scenario, will get that working in the future.

Updated README to give credit to Schaka for BuffLib which was necessary in order to detect this.